requirements.txt
keras-ocr
opencv-python==4.5.5
numpy==1.21.5
Pillow==9.0.1